# openapi.json
{"openapi":"3.0.1","info":{"title":"Payment Processing API","description":"Event-driven payment processing system.\n\n## Overview\nAccepts payment requests, processes them asynchronously through Kafka,\nand calls a payment gateway. Provides idempotency, state machine\nguarantees, and reconciliation for stuck payments.\n\n## Key concepts\n- **Idempotency**: Use the `idempotencyKey` field to safely retry\n  requests. The same key returns the same payment.\n- **Async processing**: A successful POST returns 202 with status\n  PENDING. The payment processes asynchronously and transitions to\n  SUCCESS, FAILED, or UNKNOWN.\n- **Correlation**: Pass an `X-Correlation-Id` header to trace requests\n  across logs.\n","contact":{"name":"Aryan","url":"https://github.com/aryan/payment-system"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"0.5.0"},"servers":[{"url":"http://localhost:8090","description":"Local development"}],"tags":[{"name":"Payments","description":"Create and retrieve payments"},{"name":"Health","description":"Service liveness check"}],"paths":{"/api/payments":{"post":{"tags":["Payments"],"summary":"Create a payment","description":"Creates a payment for the given order and starts asynchronous processing.\n\nThe response returns immediately with status PENDING. The payment is then\nprocessed via Kafka: the consumer calls the payment gateway and transitions\nthe payment to SUCCESS, FAILED, or UNKNOWN.\n\nIdempotency: providing the same `idempotencyKey` on retry returns the\noriginal payment without creating a duplicate.\n\nState machine: an order can have at most one non-failed payment.\nSubsequent POSTs for the same order are rejected with 422 unless the\nprevious payment failed.\n","operationId":"createPayment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePaymentRequest"}}},"required":true},"responses":{"202":{"description":"Payment accepted for processing","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"400":{"description":"Invalid request (missing fields, bad currency, etc.)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Order not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Order already has a payment in progress or paid successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/payments/{id}/approve":{"post":{"tags":["Payments"],"summary":"Approve a PENDING payment (Day 13: race-prone, intentionally)","description":"Promotes a payment from PENDING to SUCCESS.\n\n⚠️ This endpoint has a known race condition under concurrent calls\nfor the same order. It exists to demonstrate the bug Day 14's\npessimistic locking will fix. Do not rely on this in any\nproduction-like scenario.\n","operationId":"approve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"500":{"description":"Race condition — uniqueness violation under concurrency","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Payment not in PENDING or order already has SUCCESS","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Payment approved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"404":{"description":"Payment not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/payments/{id}/approve-optimistic":{"post":{"tags":["Payments"],"summary":"Approve a PENDING payment (optimistic locking, Day 15)","description":"Promotes a payment from PENDING to SUCCESS using optimistic\nlocking via @Version. No row locks are taken; conflicts are\ndetected at UPDATE time and resolved via retry.\n\nCompare to POST /api/payments/{id}/approve which uses\npessimistic locking.\n","operationId":"approveOptimistic","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"422":{"description":"Payment not PENDING, or order already has SUCCESS, or retry exhaustion","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Payment approved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"404":{"description":"Payment not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/payments/{id}":{"get":{"tags":["Payments"],"summary":"Get a payment by ID","description":"Retrieves the current state of a payment, including its status.","operationId":"getPayment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Payment found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"400":{"description":"Invalid UUID in path","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Payment not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/health":{"get":{"tags":["Health"],"summary":"Liveness check","description":"Returns UP if the service is responsive.","operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}},"components":{"schemas":{"CreatePaymentRequest":{"required":["amount","currency","idempotencyKey","orderId"],"type":"object","properties":{"orderId":{"type":"string","description":"Unique identifier of the order to pay for","format":"uuid","example":"31310001-3131-3131-3131-313131313131"},"amount":{"type":"number","description":"Payment amount in the smallest currency unit, e.g. paise for INR","example":1500.0},"currency":{"pattern":"^[A-Z]{3}$","type":"string","description":"ISO 4217 currency code","example":"INR","enum":["INR","USD","EUR"]},"idempotencyKey":{"maxLength":128,"minLength":8,"type":"string","description":"Client-chosen idempotency key. Retrying with the same key returns the original payment without creating a duplicate.","example":"client-payment-001"},"webhookUrl":{"maxLength":2048,"minLength":0,"pattern":"^(https?://).*","type":"string","description":"Optional URL where the system will POST status updates as the payment progresses","nullable":true,"example":"https://merchant.example.com/webhooks/payments"}},"description":"Request to create a new payment"},"PaymentResponse":{"type":"object","properties":{"id":{"type":"string","description":"System-generated payment ID","format":"uuid","example":"557c9e62-ef6e-4715-929d-25d4c77f5a76"},"orderId":{"type":"string","description":"Order this payment was made for","format":"uuid","example":"31310001-3131-3131-3131-313131313131"},"amount":{"type":"number","description":"Payment amount","example":1500.0},"currency":{"type":"string","description":"ISO 4217 currency code","example":"INR"},"status":{"type":"string","description":"Current status of the payment.\n- PENDING: just created, queued for processing\n- PROCESSING: consumer is actively calling the gateway\n- SUCCESS: payment completed successfully\n- FAILED: gateway rejected the payment\n- UNKNOWN: gateway response was lost; reconciliation will resolve\n","example":"SUCCESS","enum":["PENDING","PROCESSING","SUCCESS","FAILED","UNKNOWN"]},"createdAt":{"type":"string","description":"When the payment was created","format":"date-time","example":"2026-06-08T15:03:05.488774Z"}},"description":"Payment details returned after creation or status query"},"ErrorResponse":{"type":"object","properties":{"timestamp":{"type":"string","description":"When the error occurred","format":"date-time","example":"2026-06-08T15:03:05.488774Z"},"status":{"type":"integer","description":"HTTP status code","format":"int32","example":422},"error":{"type":"string","description":"Short error category","example":"Unprocessable Entity"},"message":{"type":"string","description":"Human-readable explanation","example":"Order already has a payment in progress (paymentId=..., status=PENDING)"},"path":{"type":"string","description":"The request path that produced the error","example":"/api/payments"},"details":{"type":"array","description":"Additional validation details when applicable","example":["amount: amount must be greater than 0"],"items":{"type":"string","description":"Additional validation details when applicable","example":"[\"amount: amount must be greater than 0\"]"}}},"description":"Standard error response shape"}}}}